==> 2 Analysis/Proteome/Independent/1 Single normalising data.Rmd <==
---
title: "Sample loading normalisation + IRS"
output:
---

```{r}
#######################################################################################################################################
# Independent analysis Notebook 1 - DESCRIPTION
#
# Here we take the raw data (output of Perseus without NA-filtering).
# First, we infer data for the missing sample - fortunately this was one of the middle samples, so data is inferred by taking the 
# mean of the adjacent time points.
# This is the "independent" analysis - i.e. analysing WT and CKO independently.
# Therefore we split the dataset into WT + CKO before doing the normalisation steps to each.
# We then apply sample loading normalisation to correct for small errors in pipetting - this assumes equal protein was used for
# all samples submitted for TMT labelling.
# We then apply Internal Reference Scaling as described by Phil Wilmarth - https://github.com/pwilmart/IRS_normalization
# This corrects for batch effects.
# Along the way we plot the data to make sure that we can spot obvious mistakes.
# Finally, we export the data because this is handy for plotting proteins of interest in other software!
#
# INPUTS = raw data from Perseus (.txt)
# OUTPUTS = normalised data for furtheer analysis (.csv)
#
#######################################################################################################################################
```

```{r}
#setup

# Clear your workspace
rm(list=ls())

# Load useful packages
if (!require("tidyverse")) {
  install.packages("tidyverse", dependencies = TRUE)
  library(tidyverse)
} 

if (!require("BiocManager")) {
  install.packages("BiocManager", dependencies = TRUE)
}

if (!require("edgeR")) {
  BiocManager::install("edgeR")
  library(edgeR)
} 

if (!require("psych")) {
  install.packages("psych", dependencies = TRUE)
  library(psych)
} 

# Make a list of all the .txt files in your working directory just so you can see what's going on
X = list.files(".",".txt")

# read the raw data file, and ensure "NaN" is interpreted as NA
data_start <- read.delim("CRM3962_6set_MQ1633C_prot_noFil_m8.txt", na = c("NaN"), stringsAsFactors = F)
data_start <- data_start[-1,] # Remove the first row which is just junk
data_start[, 1:60] <- sapply(data_start[,1:60], as.numeric) # Make the data numeric
missing_data <- apply(cbind(data_start[,13], data_start[,15]), 1, mean) # Infer missing values for CKO sample 12
data_start <- cbind(data_start[1:13], missing_data, data_start[15:ncol(data_start)])

```


```{r}
# Clean up majority protein IDs
proteins <- as.vector(data_start$Majority.protein.IDs)

# Create empty vector
protein_id <- vector("list", length(proteins))

# For loop to either print the name or separate them if there are several in one line
for (i in 1:length(proteins)) {
  protein_id[i] <- strsplit(proteins[i], ";")
}

# Select only the first element on each line --> this is the main protein for each line then attach it to the first column of the total proteome dataframe
protein_ids <- unlist(lapply(protein_id, '[[', 1))
data_start$uniprot <- protein_ids


# Do the same for gene names!

# Selecting the main gene name from each line
genes <- vector("list", length(data_start$Gene.names)) # Create empty vector
for (i in 1:length(data_start$Gene.names)) {
  genes[i] <- strsplit(data_start$Gene.names[i], ";")
}
gene_names <- unlist(lapply(genes, '[', 1))
data_start$main_gene <- gene_names

```


```{r}
# Separate the data into WT + CKO
WT_data_start <- cbind(data_start[,31:60], data_start[,61:ncol(data_start)])
CKO_data_start <- cbind(data_start[,1:30], data_start[,61:ncol(data_start)])

# Remove rows which contain NaN within the data columns

WT_data_start <- WT_data_start[rowSums(is.na(WT_data_start[, 1:30])) == 0,]
CKO_data_start <- CKO_data_start[rowSums(is.na(CKO_data_start[, 1:30])) == 0,]

```

```{r}
# Extract data only
WT_data_raw <- as.data.frame(WT_data_start[1:30])
CKO_data_raw <- as.data.frame(CKO_data_start[1:30])
```

```{r}
#See what raw data looks like 

boxplot(log2(WT_data_raw), col = rep(c("red", "green", "blue"), each = 10), 
        main = "WT Raw data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)

boxplot(log2(CKO_data_raw), col = rep(c("red", "green", "blue"), each = 10), 
        main = "CKO Raw data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)

# can also look at density plots (like a distribution histogram)
plotDensities(log2(WT_data_raw), col = rep(c('red', 'green', 'blue'), 6), 
              main = 'WT Raw data')

plotDensities(log2(CKO_data_raw), col = rep(c('red', 'green', 'blue'), 6), 
              main = 'CKO Raw data')

# check the column totals (per channel sums)
format(round(colSums(WT_data_raw), digits = 0), big.mark = ",")
format(round(colSums(CKO_data_raw), digits = 0), big.mark = ",")
```

```{r}
#Sample loading normalisation + IRS Normalisation (Phil Wilmarth, OHSU PSR Core, January 2018)

# separate the TMT data by experiment
exp1_raw <- WT_data_raw[c(1:10)]
exp2_raw <- WT_data_raw[c(11:20)]
exp3_raw <- WT_data_raw[c(21:30)]
exp4_raw <- CKO_data_raw[c(1:10)]
exp5_raw <- CKO_data_raw[c(11:20)]
exp6_raw <- CKO_data_raw[c(21:30)]

# figure out the global scaling value
WT_target <- mean(c(colSums(exp1_raw), colSums(exp2_raw), colSums(exp3_raw)))
CKO_target <- mean(c(colSums(exp4_raw), colSums(exp5_raw), colSums(exp6_raw)))

# do the sample loading normalization before the IRS normalization
# there is a different correction factor for each column
norm_facs <- WT_target / colSums(exp1_raw)
exp1_sl <- sweep(exp1_raw, 2, norm_facs, FUN = "*")
norm_facs <- WT_target / colSums(exp2_raw)
exp2_sl <- sweep(exp2_raw, 2, norm_facs, FUN = "*")
norm_facs <- WT_target / colSums(exp3_raw)
exp3_sl <- sweep(exp3_raw, 2, norm_facs, FUN = "*")
norm_facs <- CKO_target / colSums(exp4_raw)
exp4_sl <- sweep(exp4_raw, 2, norm_facs, FUN = "*")
norm_facs <- CKO_target / colSums(exp5_raw)
exp5_sl <- sweep(exp5_raw, 2, norm_facs, FUN = "*")
norm_facs <- CKO_target / colSums(exp6_raw)
exp6_sl <- sweep(exp6_raw, 2, norm_facs, FUN = "*")

# make a pre-IRS data frame after sample loading norms
WT_data_sl <- cbind(exp1_sl, exp2_sl, exp3_sl)
CKO_data_sl <- cbind(exp4_sl, exp5_sl, exp6_sl)

# see what the SL normalised data look like
boxplot(log2(WT_data_sl), col = rep(c("red", "green", "blue"), each = 10), 
        main = "WT Sample loading (SL) normalized data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)
boxplot(log2(CKO_data_sl), col = rep(c("red", "green", "blue"), each = 10), 
        main = "CKO Sample loading (SL) normalized data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)

# can also look at density plots (like a distribution histogram)    
plotDensities(log2(WT_data_sl), col = rep(c("red", "green", "blue"), 10), main = "SL data")
plotDensities(log2(CKO_data_sl), col = rep(c("red", "green", "blue"), 10), main = "SL data")

# check the columnn totals
format(round(colSums(WT_data_sl), digits = 0), big.mark = ",")
format(round(colSums(CKO_data_sl), digits = 0), big.mark = ",")
```

```{r}
# make working frame with row sums from each frame
WT_irs <- tibble(apply(exp1_sl[,9:10], 1, mean), apply(exp2_sl[,9:10], 1, mean), apply(exp3_sl[,9:10], 1, mean))
CKO_irs <- tibble(apply(exp4_sl[,9:10], 1, mean), apply(exp5_sl[,9:10], 1, mean), apply(exp6_sl[,9:10], 1, mean))

colnames(WT_irs) <- c("mean1", "mean2", "mean3")
colnames(CKO_irs) <- c("mean4", "mean5", "mean6")

# get the average intensity for each protein
WT_irs$average <- apply(WT_irs, 1, mean)
CKO_irs$average <- apply(CKO_irs, 1, mean)

# compute the scaling factor vectors
WT_irs$fac1 <- WT_irs$average / WT_irs$mean1
WT_irs$fac2 <- WT_irs$average / WT_irs$mean2
WT_irs$fac3 <- WT_irs$average / WT_irs$mean3
CKO_irs$fac4 <- CKO_irs$average / CKO_irs$mean4
CKO_irs$fac5 <- CKO_irs$average / CKO_irs$mean5
CKO_irs$fac6 <- CKO_irs$average / CKO_irs$mean6

# make new data frame with normalized data
WT_data_irs <- exp1_sl * WT_irs$fac1
WT_data_irs <- cbind(WT_data_irs, exp2_sl * WT_irs$fac2)
WT_data_irs <- cbind(WT_data_irs, exp3_sl * WT_irs$fac3)

CKO_data_irs <- exp4_sl * CKO_irs$fac4
CKO_data_irs <- cbind(CKO_data_irs, exp5_sl * CKO_irs$fac5)
CKO_data_irs <- cbind(CKO_data_irs, exp6_sl * CKO_irs$fac6)

# see what the IRS data look like
boxplot(log2(WT_data_irs), col = rep(c("red", "green", "blue"), each = 10), 
        main = "WT Internal Reference Scaling (IRS) normalized data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)
boxplot(log2(CKO_data_irs), col = rep(c("red", "green", "blue"), each = 10), 
        main = "CKO Internal Reference Scaling (IRS) normalized data: \nExp1 (red), Exp2 (green), Exp3 (blue)",
        xlab = 'TMT Sample', ylab = 'log2 of Intensity', notch = TRUE)

# can also look at density plots (like a distribution histogram)    
plotDensities(log2(WT_data_irs), col = rep(c("red", "green", "blue"), 10), main = "IRS data")
plotDensities(log2(CKO_data_irs), col = rep(c("red", "green", "blue"), 10), main = "IRS data")

# check column totals
format(round(colSums(WT_data_irs), digits = 0), big.mark = ",")
format(round(colSums(CKO_data_irs), digits = 0), big.mark = ",")

```

```{r}
# see how things cluster now that we have nice boxplots and density plots
plotMDS(log2(WT_data_irs), col = rep(c("red", "green", "blue"), each = 10), 
        main = "SL/TMM clusters group by TMT experiment")
plotMDS(log2(CKO_data_irs), col = rep(c("red", "green", "blue"), each = 10), 
        main = "SL/TMM clusters group by TMT experiment")
```

```{r}
#Export normalised data of both genotypes
# Remove pools from the dataframe of IRS-normalised data
# Add uniprot ID, protein names, gene names

WT_data <- WT_data_irs %>% dplyr::select(-contains("P"))
WT_data <- cbind(WT_data_start$uniprot, WT_data_start$main_gene, WT_data_start$Protein.names, WT_data)
colnames(WT_data)[1:3] <- c("uniprot", "main_gene", "protein_name")

CKO_data <- CKO_data_irs %>% dplyr::select(-contains("P"))
CKO_data <- cbind(CKO_data_start$uniprot, CKO_data_start$main_gene, CKO_data_start$Protein.names, CKO_data)
colnames(CKO_data)[1:3] <- c("uniprot", "main_gene", "protein_name")

write.csv(WT_data, "All_WT_normalised_data.csv", row.names = F)
write.csv(CKO_data, "All_CKO_normalised_data.csv", row.names = F)
```


```{r}
# Prepare .txt files for eJTK

# Separate into WT + CKO with only uniprot IDs
WTeJTKcomb <- cbind(WT_data$uniprot, dplyr::select(WT_data, contains("WT")))
CKOeJTKcomb <- cbind(CKO_data$uniprot, CKO_data[,4:ncol(CKO_data)])

# Colnames need to be: "#" for IDs, then CT or ZT preceding timepoint number
colnames(WTeJTKcomb) <- c("ID", paste(rep("CT", 24), seq(0, 69, 3), sep=""))
colnames(CKOeJTKcomb) <- c("ID", paste(rep("CT", 24), seq(0, 69, 3), sep=""))

# Files need to be tab-separated
write.table(WTeJTKcomb, file = "WTeJTKcomb-single.txt", sep = "\t", row.names = F, col.names = T, quote = F)
write.table(CKOeJTKcomb, file = "CKOeJTKcomb-single.txt", sep = "\t", row.names = F, col.names = T, quote = F)

```
