==> 2 Analysis/Proteome/Independent/11 Calculating fold change in abundance over time.Rmd <==
---
title: "Calculating fold-change in abundance over time"
output:
---

```{r}
#######################################################################################################################################
# Independent analysis Notebook 11 - DESCRIPTION
#
# Here we take all normalised abundance data and calculate the fold-change between maximum and minimum values.
# We then calculate the Cohen D statistic for effect size, as recommended by Reviewer 1.
# However this is not valid because our data is not normally distributed.
# We go on to plot probability density, and apply the Kolmogorov-Smirnov test, clearly illustrating a difference in distribution.
#
# INPUTS = all normalised data (.csv)
# OUTPUTS = Cohen D statistic, probability density plots, KS test
# 
#######################################################################################################################################
```

```{r}
# Setup
# Clear your workspace
rm(list=ls())

# Load useful packages

if (!require("effsize")) {
  install.packages("effsize", dependencies = TRUE)
  library(effsize)
} # this is for the Cohen D statistic

if (!require("tidyverse")) {
  install.packages("tidyverse", dependencies = TRUE)
  library(tidyverse)
} 

if (!require("gridExtra")) {
  install.packages("gridExtra", dependencies = TRUE)
  library(gridExtra)
}

# Make a list of all the .csv files in your working directory just so you can see what's going on
X = list.files(".",".csv")

# read the normalised data files
WT_data_start <- read.csv("All_WT_normalised_data.csv")
CKO_data_start <- read.csv("All_CKO_normalised_data.csv")

```

```{r}
# Calculate fold changes

WT <- apply(WT_data_start[, 4:ncol(WT_data_start)], 1, function(x){max(x)/min(x)})
CKO <- apply(CKO_data_start[, 4:ncol(CKO_data_start)], 1, function(x){max(x)/min(x)})

```

```{r}
# Calculate Cohen D (NB: our data is not normally distributed so this may not be valid)

cohen.d(CKO, WT)

```

```{r}
# Kolmogorov-Smirnov test for difference in distribution

st <- ks.test(WT, CKO)

# Plot probability distributions
# set colours
colours <- c("WT" = "blue", "CKO" = "red")

pd <- ggplot() +
  geom_density(aes(x = WT, fill = "WT"),
               color = "black", alpha = 0.5) +
  geom_density(aes(x = CKO, fill = "CKO"),
               color = "black", alpha = 0.5) +
  scale_fill_manual(name = "", values = colours) +
  xlim(1, 3) +
  labs(subtitle = paste("KS test p-val: ", format(round(st$p.value,4))), y="Probability density", x="Fold Change") +
  theme(panel.border = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.background = element_blank(),
        axis.ticks = element_line(colour = "black", size=1),
        axis.line = element_line(colour = "black", size=1),
        axis.text=element_text(colour="black", face="bold", size=11),
        axis.title.x = element_text(colour= "black", face="bold", size=12),
        axis.title.y = element_text(colour= "black", face="bold", size=12),
        legend.background = element_blank(),
        legend.key = element_blank(),
        legend.text = element_text(colour="black", face="bold", size=11)
        )

# Plot cumulative distribution for reference 

cp <- ggplot() +
  geom_step(aes(x=WT, colour = "WT"), stat="ecdf") + 
  geom_step(aes(x=CKO, colour = "CKO"), stat="ecdf") + 
  scale_color_manual(name = "", values = colours) +
  labs(subtitle = paste("KS test p-val: ", format(round(st$p.value,4))), y="Cumulative probability", x="Fold Change") +
  xlim(1,3) +
  theme(panel.border = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.background = element_blank(),
        axis.ticks = element_line(colour = "black"),
        axis.line = element_line(colour = "black"),
        axis.text=element_text(colour="black"),
        legend.background = element_blank(),
        legend.key = element_blank()
        )

grid.arrange(pd, cp, nrow=2)

ggsave(plot=pd, "Proteome prob densities.pdf", width=7, height = 5)

```
